NBA Free Throws Data

A dataframe of NBA free throw data as supplied by Mango Solutions

free_throws

Format

A data frame with 618,019 rows and 11 variables.

end_result

Character, final score of the game.

game

Character, identifies the two teams playing

game_id

Numeric game id.

period

Numeric, which period was the play made in.

play

Character string outlining the play

player

Character, the name of the player

playoffs

Character string, type of game

score

Character string identifying the final score

season

Character identifying the season the game was played in.

shot_made

Binary, was a shot made

time

Time at which a play occurred