Sample from a Probability Distribution

Function brief: Write a function that will draw random numbers from a given distribution. The function should take one argument for the number of samples and a second argument which specifies the distribution (Normal, Poisson or Binomial). The function should be able to handle additional parameters depending on the distribution chosen, e.g., a lambda for Poisson samples.

sample_dist(n = NULL, dist = NULL, ...)

Arguments

n
Numeric (or integer), the number of samples to take.

dist
A function which samples from a probability distribution. The first arguement must be the number of samples, additional arguements can be specified with .... Possible options are rnorm (normal), rpois (Poisson), and rbinom (Binomial).

...
Additional arguements to pass to the specified dist function.

Value

A vector of samples from the specified distribution.

Examples

## Sample a single time from the normal distribution
sample_dist(1, dist = rnorm, mean = 0, sd = 1)

#> [1] -1.400044
#> attr(,"class")
#> [1] "sample_dist" "numeric"    

## Sample multiple times from the poisson distribution
sample_dist(100, dist = rpois, lambda = 2)

#>   [1] 2 1 0 2 2 1 3 3 4 1 0 1 1 1 1 0 1 5 1 3 3 1 5 3 0 2 3 3 0 1 1 2 2 2 3 5 1
#>  [38] 1 3 2 2 2 0 3 3 6 5 1 2 1 1 2 2 3 1 3 0 1 3 1 2 1 2 1 5 2 2 1 2 1 0 2 1 0
#>  [75] 1 2 3 1 2 2 1 1 2 4 1 1 2 0 7 1 2 3 3 1 2 3 0 2 3 3
#> attr(,"class")
#> [1] "sample_dist" "integer"